// src/app/model/action-name.enum.ts
export enum ActionName {
  UPDATE_CURRENCIES = 'UPDATE_CURRENCIES',
  SELECT_CRYPTOCURRENCY = 'SELECT_CRYPTOCURRENCY',
  CHOOSE_CRYPTOCURRENCY = 'CHOOSE_CRYPTOCURRENCY',
  SELECT_FIAT_CURRENCY = 'SELECT_FIAT_CURRENCY',
  UPDATE_FIAT_CURRENCY = 'UPDATE_FIAT_CURRENCY',
  RELOAD_CRYPTOCURRENCIES = 'RELOAD_CRYPTOCURRENCIES',
  LOAD_CRYPTOCURRENCIES = 'LOAD_CRYPTOCURRENCIES',
  RELOAD_CRYPTOCURRENCY_DETAILS = 'RELOAD_CRYPTOCURRENCY_DETAILS',
  TURN_OFF_MULTIPLE_CONVERSIONS = 'TURN_OFF_MULTIPLE_CONVERSIONS',
}